+++ nucypher-core-wasm/tests/wasm.rs
@@ -1,5 +1,3 @@
-use nucypher_core_wasm::*;
-
 use ferveo::bindings_wasm::{ferveo_encrypt, DkgPublicKey, FerveoVariant, Keypair};
 use umbral_pre::bindings_wasm::{
     generate_kfrags, reencrypt, Capsule, RecoverableSignature, SecretKey, Signer,
@@ -8,6 +6,8 @@ use umbral_pre::bindings_wasm::{
 use wasm_bindgen::prelude::*;
 use wasm_bindgen::JsCast;
 use wasm_bindgen_test::*;
+
+use nucypher_core_wasm::*;
 
 //
 // Test utilities
@@ -699,14 +699,23 @@ fn threshold_decryption_request() {
     let context: JsValue = Some(Context::new("{'user': 'context'}")).into();
 
     let dkg_pk = DkgPublicKey::random();
+
+    let auth_data =
+        AuthenticatedData::new(&dkg_pk, &conditions_js.unchecked_into::<OptionConditions>())
+            .unwrap();
+
+    let authorization = b"we_dont_need_no_stinking_badges";
+    let acp = AccessControlPolicy::new(&auth_data, authorization).unwrap();
+
     let message = "my-message".as_bytes();
-    let ciphertext = ferveo_encrypt(message, conditions.as_bytes(), &dkg_pk).unwrap();
+    let ciphertext = ferveo_encrypt(message, &acp.aad(), &dkg_pk).unwrap();
+    let ciphertext_header = ciphertext.header().unwrap();
 
     let request = ThresholdDecryptionRequest::new(
         ritual_id,
         &FerveoVariant::simple(),
-        &ciphertext,
-        &conditions_js.unchecked_into::<OptionConditions>(),
+        &ciphertext_header,
+        &acp,
         &context.unchecked_into::<OptionContext>(),
     )
     .unwrap();
@@ -786,4 +795,112 @@ fn threshold_decryption_response() {
     assert!(encrypted_response_from_bytes
         .decrypt(&random_shared_secret)
         .is_err());
+}
+
+#[wasm_bindgen_test]
+fn authenticated_data() {
+    let dkg_pk = DkgPublicKey::random();
+
+    let conditions = "{'some': 'condition'}";
+    let conditions_js: JsValue = Some(Conditions::new(conditions)).into();
+
+    let auth_data =
+        AuthenticatedData::new(&dkg_pk, &conditions_js.unchecked_into::<OptionConditions>())
+            .unwrap();
+
+    assert_eq!(
+        auth_data.public_key().to_bytes().unwrap(),
+        dkg_pk.to_bytes().unwrap()
+    );
+    assert_eq!(auth_data.conditions().unwrap().to_string(), conditions);
+
+    let mut expected_aad = dkg_pk.to_bytes().unwrap().to_vec();
+    expected_aad.extend(conditions.as_bytes());
+
+    assert_eq!(auth_data.aad(), expected_aad.into_boxed_slice());
+
+    // mimic serialization/deserialization over the wire
+    let serialized_auth_data = auth_data.to_bytes();
+    let deserialized_auth_data = AuthenticatedData::from_bytes(&serialized_auth_data).unwrap();
+    assert_eq!(
+        deserialized_auth_data.public_key().to_bytes().unwrap(),
+        dkg_pk.to_bytes().unwrap()
+    );
+    assert_eq!(
+        deserialized_auth_data.conditions().unwrap().to_string(),
+        conditions,
+    );
+}
+
+#[wasm_bindgen_test]
+fn access_control_policy() {
+    let dkg_pk = DkgPublicKey::random();
+
+    let conditions = "{'some': 'condition'}";
+    let conditions_js: JsValue = Some(Conditions::new(conditions)).into();
+
+    let auth_data =
+        AuthenticatedData::new(&dkg_pk, &conditions_js.unchecked_into::<OptionConditions>())
+            .unwrap();
+
+    let authorization = b"we_dont_need_no_stinking_badges";
+    let acp = AccessControlPolicy::new(&auth_data, authorization).unwrap();
+
+    assert_eq!(
+        dkg_pk.to_bytes().unwrap(),
+        acp.public_key().to_bytes().unwrap()
+    );
+    assert_eq!(
+        authorization.to_vec().into_boxed_slice(),
+        acp.authorization()
+    );
+    assert_eq!(conditions, acp.conditions().unwrap().to_string());
+
+    // mimic serialization/deserialization over the wire
+    let serialized_acp = acp.to_bytes();
+    let deserialized_acp = AccessControlPolicy::from_bytes(&serialized_acp).unwrap();
+    assert_eq!(
+        dkg_pk.to_bytes().unwrap(),
+        deserialized_acp.public_key().to_bytes().unwrap()
+    );
+    assert_eq!(
+        authorization.to_vec().into_boxed_slice(),
+        deserialized_acp.authorization()
+    );
+    assert_eq!(
+        conditions,
+        deserialized_acp.conditions().unwrap().to_string()
+    );
+
+    // check aad; expected acp and auth_data acps to be the same
+    assert_eq!(deserialized_acp.aad(), auth_data.aad());
+}
+
+#[wasm_bindgen_test]
+fn threshold_message_kit() {
+    let conditions = "{'some': 'condition'}";
+    let conditions_js: JsValue = Some(Conditions::new(conditions)).into();
+
+    let dkg_pk = DkgPublicKey::random();
+
+    let auth_data =
+        AuthenticatedData::new(&dkg_pk, &conditions_js.unchecked_into::<OptionConditions>())
+            .unwrap();
+
+    let authorization = b"we_dont_need_no_stinking_badges";
+    let acp = AccessControlPolicy::new(&auth_data, authorization).unwrap();
+
+    let data = "The Tyranny of Merit".as_bytes();
+    let ciphertext = ferveo_encrypt(data, &acp.aad(), &dkg_pk).unwrap();
+
+    let tmk = ThresholdMessageKit::new(&ciphertext, &acp);
+
+    // mimic serialization/deserialization over the wire
+    let serialized_tmk = tmk.to_bytes();
+    let deserialized_tmk = ThresholdMessageKit::from_bytes(&serialized_tmk).unwrap();
+    assert_eq!(
+        ciphertext.header().unwrap(),
+        deserialized_tmk.ciphertext_header().unwrap()
+    );
+    assert_eq!(acp, deserialized_tmk.acp());
 }